Config relocated
